refactor: InputTheme props 추가 및 styles 분리
